int it is

+++ UPS_Lite.py
@@ -22,7 +22,7 @@ class UPS():
                         prev_capacity = tmpfile.read()
                 except FileNotFoundError:
                         prev_capacity = "1000"
-                return float(prev_capacity)
+                return int(prev_capacity)
 
 
         def read_voltage(self):
@@ -48,7 +48,7 @@ class UPS():
                 tmpfile.write(str(capacity))
                 tmpfile.close
 
-                return float(capacity)
+                return int(capacity)
 
         def is_battery_full(self,capacity):
